Split set structs into separate files
Part of #2
